+++ packages/components/core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.27.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-core@0.26.2...@synerise/ds-core@0.27.0) (2021-02-05)
+
+
+### Features
+
+* **table:** add selection limit warning message ([71433b4](https://github.com/synerise/synerise-design/commit/71433b406924bc7ec098f71d6e36d58ab0f6ad82))
+
+
+
+
+
 ## [0.26.2](https://github.com/synerise/synerise-design/compare/@synerise/ds-core@0.26.1...@synerise/ds-core@0.26.2) (2021-02-02)
 
 **Note:** Version bump only for package @synerise/ds-core