+++ packages/components/input/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.20](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.16.19...@synerise/ds-input@0.16.20) (2021-02-05)
+
+**Note:** Version bump only for package @synerise/ds-input
+
+
+
+
+
 ## [0.16.19](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.16.18...@synerise/ds-input@0.16.19) (2021-02-02)
 
 **Note:** Version bump only for package @synerise/ds-input