+++ packages/components/manageable-list/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.5](https://github.com/Synerise/synerise-design/compare/@synerise/ds-manageable-list@0.16.4...@synerise/ds-manageable-list@0.16.5) (2021-02-05)
+
+**Note:** Version bump only for package @synerise/ds-manageable-list
+
+
+
+
+
 ## [0.16.4](https://github.com/Synerise/synerise-design/compare/@synerise/ds-manageable-list@0.16.3...@synerise/ds-manageable-list@0.16.4) (2021-02-02)
 
 **Note:** Version bump only for package @synerise/ds-manageable-list