+++ packages/components/sidebar-object/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.24](https://github.com/Synerise/synerise-design/compare/@synerise/ds-sidebar-object@0.4.23...@synerise/ds-sidebar-object@0.4.24) (2021-02-05)
+
+**Note:** Version bump only for package @synerise/ds-sidebar-object
+
+
+
+
+
 ## [0.4.23](https://github.com/Synerise/synerise-design/compare/@synerise/ds-sidebar-object@0.4.22...@synerise/ds-sidebar-object@0.4.23) (2021-02-02)
 
 **Note:** Version bump only for package @synerise/ds-sidebar-object