+++ packages/components/slider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.9.2...@synerise/ds-slider@0.9.3) (2021-02-05)
+
+**Note:** Version bump only for package @synerise/ds-slider
+
+
+
+
+
 ## [0.9.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.9.1...@synerise/ds-slider@0.9.2) (2021-02-02)
 
 **Note:** Version bump only for package @synerise/ds-slider